feat(release): Generate yaml with friendlyApiName key for reference docs (#78)
* feat(release): Generate yaml with friendlyApiName key for reference docs

This adds a new argument to the build-rad tool to accept the friendly API name of a gem. This is then added into the yard template for reference pages.

* feat(release): Generate yaml with friendlyApiName key for reference docs

This adds a new argument to the build-rad tool to accept the friendly API name of a gem. This is then added into the yard template for reference pages.

* feat(release): Generate yaml with friendlyApiName key for reference docs

This adds a new argument to the build-rad tool to accept the friendly API name of a gem. This is then added into the yard template for reference pages.

* feat(release): Generate yaml with friendlyApiName key for reference docs

This adds a new argument to the build-rad tool to accept the friendly API name of a gem. This is then added into the yard template for reference pages.